+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Subscription notification message once successfully subscribed ([#104](https://github.com/commanded/eventstore/pull/104)).
 - Transient subscriptions ([#105](https://github.com/commanded/eventstore/pull/105)).
 - Use a single PostgreSQL connection for all subscriptions ([#106](https://github.com/commanded/eventstore/pull/106)).
+- Transient subscription event mapping function ([#108](https://github.com/commanded/eventstore/pull/108)).
 
 ## v0.13.2